+++ src/schemathesis/runner/impl/solo.py
@@ -6,7 +6,7 @@ import attr
 from ...models import TestResultSet
 from ...utils import get_requests_auth
 from .. import events
-from .core import BaseRunner, get_session, network_test, run_test, wsgi_test
+from .core import BaseRunner, get_session, network_test, wsgi_test
 
 
 @attr.s(slots=True)  # pragma: no mutate
@@ -16,38 +16,34 @@ class SingleThreadRunner(BaseRunner):
     def _execute(self, results: TestResultSet) -> Generator[events.ExecutionEvent, None, None]:
         auth = get_requests_auth(self.auth, self.auth_type)
         with get_session(auth) as session:
-            for endpoint, test in self.schema.get_all_tests(network_test, self.hypothesis_settings, self.seed):
-                for event in run_test(
-                    endpoint,
-                    test,
-                    self.checks,
-                    self.targets,
-                    results,
-                    session=session,
-                    headers=self.headers,
-                    request_timeout=self.request_timeout,
-                    store_interactions=self.store_interactions,
-                ):
-                    yield event
-                    if isinstance(event, events.Interrupted):
-                        return
+            yield from self._run_tests(
+                self.schema.get_all_tests,
+                network_test,
+                self.hypothesis_settings,
+                self.seed,
+                checks=self.checks,
+                targets=self.targets,
+                results=results,
+                session=session,
+                headers=self.headers,
+                request_timeout=self.request_timeout,
+                store_interactions=self.store_interactions,
+            )
 
 
 @attr.s(slots=True)  # pragma: no mutate
 class SingleThreadWSGIRunner(SingleThreadRunner):
     def _execute(self, results: TestResultSet) -> Generator[events.ExecutionEvent, None, None]:
-        for endpoint, test in self.schema.get_all_tests(wsgi_test, self.hypothesis_settings, self.seed):
-            for event in run_test(
-                endpoint,
-                test,
-                self.checks,
-                self.targets,
-                results,
-                auth=self.auth,
-                auth_type=self.auth_type,
-                headers=self.headers,
-                store_interactions=self.store_interactions,
-            ):
-                yield event
-                if isinstance(event, events.Interrupted):
-                    return
+        yield from self._run_tests(
+            self.schema.get_all_tests,
+            wsgi_test,
+            self.hypothesis_settings,
+            self.seed,
+            checks=self.checks,
+            targets=self.targets,
+            results=results,
+            auth=self.auth,
+            auth_type=self.auth_type,
+            headers=self.headers,
+            store_interactions=self.store_interactions,
+        )